backend/requirements.txt
asgiref==3.6.0
attrs==23.1.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
channels==4.0.0
charset-normalizer==3.1.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
daphne==4.0.0
defusedxml==0.7.1
distlib==0.3.6
Django==4.2.1
django-cors-headers==3.14.0
django-debug-toolbar==4.0.0
django-filter==23.2
django-phonenumber-field==7.1.0
django-querycount==0.8.3
django-silk==5.0.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
filelock==3.12.0
flake8==6.0.0
gprof2dot==2022.7.29
gunicorn==20.1.0
hyperlink==21.0.0
identify==2.5.24
idna==3.4
incremental==22.10.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
nodeenv==1.8.0
oauthlib==3.2.2
packaging==23.1
phonenumbers==8.13.11
phonenumberslite==8.13.11
Pillow==9.5.0
platformdirs==3.5.1
pre-commit==3.3.2
psycopg2-binary==2.9.5
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.6.0
pyOpenSSL==23.1.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
service-identity==21.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.23.0
zope.interface==6.0
